# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mrprototcol",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://mrprotocol.com/api/v1",
            "description": "Base URL"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Base URL"
        }
    ],
    "paths": {
...
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send password reset link",
                "description": "Send password reset link.",
                "operationId": "febbd99ec02533a62072c943c8864557",
                "requestBody": {
                    "description": "Request body containing user email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful password reset link sending",
                                            "type": "string",
                                            "example": "Password reset link sent successfully"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user with the provided information.",
                "operationId": "f024dbf1273b8a27c30d428e9c1fb20d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "string",
                                            "example": "201"
                                        },
                                        "access-token": {
                                            "type": "string",
                                            "example": "your_access_token"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/auth/email/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify user email",
                "operationId": "c6dcfe1647a381d8cee802ca0d744027",
                "requestBody": {
                    "description": "Request body containing the verification token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Verification token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful user verification",
                                            "type": "string",
                                            "example": "User verified successfully"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating failure to verify email",
                                            "type": "string",
                                            "example": "Failed to verify email"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating invalid verification token",
                                            "type": "string",
                                            "example": "Invalid verification token"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "title": "User Resource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "user"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "500": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Oops something went wrong"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 500
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Validation Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Validation error"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "field_name": [
                                            "The field_name field is required."
                                        ]
                                    }
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 422
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}